+++ backend/main.py
@@ -60,9 +60,11 @@ class ScreenEndpoint(ElementEndpointBase):
     _session_cls = Session
     _element = Screen
     _owner_attr = 'user_id'
-    _other_readable = list(['id', 'desc', 'template_id', 'user_id', 'duration', 'till', 'repeat', 'loop', 'variables', 'locked', 'displayed', 'default', 'key'])
-    _other_createable = list(['desc', 'template_id', 'user_id', 'duration', 'till', 'repeat', 'loop', 'variables'])
-    _all_readable = list(['id', 'duration', 'till', 'repeat', 'loop', 'variables', 'key'])
+    _other_readable = [
+        'id', 'desc', 'template_id', 'user_id', 'header', 'duration', 'till', 'repeat', 'loop', 'variables', 'locked', 'displayed', 'default', 'key'
+    ]
+    _other_createable = list(['desc', 'template_id', 'user_id', 'header', 'duration', 'till', 'repeat', 'loop', 'variables'])
+    _all_readable = list(['id', 'header', 'duration', 'till', 'repeat', 'loop', 'variables', 'key'])
 
 
 class ChallongeTournamentEndpoint(ElementEndpointBase):