+++ frontend/src/app/components/screens/challonge-parallel-tournaments/challonge-parallel-tournaments.component.ts
@@ -22,6 +22,7 @@ import { faBoltLightning } from '@fortawesome/free-solid-svg-icons';
 export class ChallongeParallelTournamentsComponent {
     faBoltLightning = faBoltLightning;
     isActive = input.required<boolean>();
+    header = input.required<string>();
     variables = input.required<any>();
     finished = output<null>();